+++ inc/core/artist-platform-post-types.php
@@ -60,7 +60,7 @@ function extrachill_register_artist_profile_cpt() {
         'show_in_admin_bar'     => true,
         'show_in_nav_menus'     => true,
         'can_export'            => true,
-        'has_archive'           => false,
+        'has_archive'           => 'artists',
         'exclude_from_search'   => false,
         'publicly_queryable'    => true,
         'rewrite'               => false,